requirements.txt
numpy
obspy==1.1.1
PyYAML==5.1
pyflex
pytest==4.5.0
git+git://github.com/lsawade/spaceweight.git@master
git+git://github.com/lsawade/pycmt3d.git@devel
